Fixed SubGraph by adding proper order between Parameter ... Result
Fixed BuildExecutableSequence_MatMul_Success after update implementation of OperationBufferExtractor

Fixed OperationBuffer Extractor for SubGraph-s

Apply clang-format

Fixed TensorIterator on decoder_iter.xml

Updated tests for TensorIterator to catch the error in Tacotron2 decoder_iter.xml

Added method for cloning ConcatOptimized

Changed FullyConnected transformation to be based on ngraph::pass::MatcherPass instead of ngraph::pass::FunctionPass and fixed wrong fusing with constant of wrong shape

Fix for FullyConnectedOp that previously divided wrong number, switched on dividing shape sizes

Add test that reproduce issue with 2 MatMul and FullyConnected

Initial support for BiriderciotalLSTMSequence

Successfully created LSTMSequence with BIDIRECTIONAL mode

Reworked Bidirectional transformation that it compose all needed transformations inside

Fixed Result operation name

Decreased number of copies for FullyConnected transformation

Create transformation that fuse Tanh to FusedConvolution

+++ modules/cuda_plugin/src/ops/fully_connected.hpp
@@ -28,7 +28,7 @@ public:
 private:
     MatMulOp matmul_op_;
     size_t bias_size_ = 0;
-    int batch_bias_count_ = 0;
+    size_t batch_bias_count_ = 0;
 };
 
 }  // namespace CUDAPlugin
